VS port ACL
I'm porting this thing to msvc (although I'm not yet sure if I'll maintain it this way).
I initially did it to get access to a profiler, but now that I've found out and solved
the problem without it, I'm not sure why I'd need it at all. But I'll keep what I've done so far